+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Added
+
+- Added back `_EXECUTOR_PLUGIN_DEFAULTS` to ssh plugin
+
 ## [0.7.0] - 2022-08-19
 
 ### Changed